graph moved to separeted component (problems with pushing prices)
